+++ src/functionalTest/groovy/com/autonomousapps/jvm/AbstractJvmSpec.groovy
@@ -7,8 +7,6 @@ import com.autonomousapps.kit.GradleProject
 
 abstract class AbstractJvmSpec extends AbstractFunctionalSpec {
 
-  protected GradleProject gradleProject = null
-
   /**
    * Set to `false` in a concrete class temporarily if you want to inspect the build output.
    */
@@ -18,13 +16,5 @@ abstract class AbstractJvmSpec extends AbstractFunctionalSpec {
     if (gradleProject != null && shouldClean) {
       clean(gradleProject.rootDir)
     }
-  }
-
-  List<Advice> actualAdvice() {
-    return AdviceHelper.actualAdviceForFirstSubproject(gradleProject)
-  }
-
-  String actualAdviceConsole() {
-    return AdviceHelper.actualConsoleAdvice(gradleProject)
   }
 }