+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Dependency Analysis Plugin Changelog
 
+# Version 0.47.0 (unreleased) 
+* [New] The concept of dependency "facades" has been formalized as "logical dependencies", aka
+"bundles". `dependencyAnalysis.setFacadeGroups()` is now deprecated and is a no-op. Users should 
+instead use `dependencyAnalysis.dependencies {}` to specify their bundles. Finally, with this 
+change, the Kotlin stdlib family is no longer considered a bundle by default.
+
 # Version 0.46.0
 * [New] New `registerPostProcessingTask()` method on the extension to register a custom task to 
 execute with the `ComprehensiveAdvice` produced by the project as an input.